Fixed an issue with post-processing the autoscan result
To give some background: When the maven-autotools-plugin doesn't
find a configure.{ac,in} file under src/main/autotools, it begins
with running the autoscan command to generate a preliminary template
instead, named 'configure.scan'.  After that, a script called
'autoscan-post' performs some minimal steps to turn that template
into a valid 'configure.ac'.

Prior to this commit, a simplistic sed script was used for post-
processing the 'configure.scan' template.  In some situations, the
autotools chain was run a second time, so that for example the
Makefile already existed.  Final post-processing then caused the
M4 macro call 'AC_CONFIG_FILES([Makefile])' to appear twice in
the 'configure.ac', which in turn caused the automake command to
fail.

Now the post-processing of the 'configure.scan' template is done
by a slightly more savvy awk script, which avoids inserting the
macro call repeatedly.