Improved loading of a native library in the archetype code
The archetype for a shared library now provides some code to determine
the system architecture and the operating system as well as the file
extender of shared libraries.  All this is necessary to load a native
library in the method 'loadNativeLibrary'.

It should be noted that there's some duplication of code originally
found in 4367d3e9d2.  Although this violates the DRY (don't repeat
yourself) principle, we refrain from introducing an additional
dependency in the archetype code.